Support exporting pages from a specific namespace
Add the -s, --namespace option to specify a specific namespace ID, as
opposed to always exporting the main namespace (but this remains the
default behavior).

If --category is specified, this option acts as a filter rather than a
selector; see https://www.mediawiki.org/wiki/API:Categorymembers for
allowed filters.